@@ -23,13 +23,13 @@
     </modules>
 
     <properties>
-        <java.version>1.8</java.version>
-        <spring.cloud.alibaba.version>2021.0.6.1</spring.cloud.alibaba.version>
-        <spring.cloud.version>2021.0.8</spring.cloud.version>
-        <spring.boot.version>2.6.15</spring.boot.version>
-        <mybatis.plus.version>3.3.0</mybatis.plus.version>
-        <redisson.version>3.32.0</redisson.version>
-        <hutool.version>5.5.8</hutool.version>
+        <java.version>17</java.version>
+        <spring.cloud.alibaba.version>2023.0.1.0</spring.cloud.alibaba.version>
+        <spring.cloud.version>2023.0.1</spring.cloud.version>
+        <spring.boot.version>3.2.4</spring.boot.version>
+        <mybatis.plus.version>3.5.5</mybatis.plus.version>
+        <hutool.version>5.8.23</hutool.version>
+
         <lombok.version>1.18.34</lombok.version>
         <jackson.version>2.11.4</jackson.version>
         <mysql.connector.java>8.0.33</mysql.connector.java>
@@ -37,7 +37,8 @@
         <commons.lang3.version>3.1</commons.lang3.version>
         <commons-pool2.version>2.9.0</commons-pool2.version>
         <mapstruct.version>1.5.5.Final</mapstruct.version>
-        <rocketmq-springboot.version>2.2.0</rocketmq-springboot.version>
+        <rocketmq-springboot.version>2.2.3</rocketmq-springboot.version>
+        <rocketmq-client.version>4.9.4</rocketmq-client.version>
         <maven-compiler-plugin.version>3.8.1</maven-compiler-plugin.version>
     </properties>
 
@@ -63,6 +64,12 @@
                 <version>${spring.cloud.alibaba.version}</version>
                 <type>pom</type>
                 <scope>import</scope>
+                <exclusions>
+                    <exclusion>
+                        <groupId>org.apache.rocketmq</groupId>
+                        <artifactId>rocketmq-client</artifactId>
+                    </exclusion>
+                </exclusions>
             </dependency>
 
             <!--    集成三方框架        -->
@@ -72,8 +79,13 @@
                 <version>${rocketmq-springboot.version}</version>
             </dependency>
             <dependency>
+                <groupId>org.apache.rocketmq</groupId>
+                <artifactId>rocketmq-client</artifactId>
+                <version>${rocketmq-client.version}</version> <!-- 确保与spring-boot-starter版本兼容 -->
+            </dependency>
+            <dependency>
                 <groupId>com.baomidou</groupId>
-                <artifactId>mybatis-plus-boot-starter</artifactId>
+                <artifactId>mybatis-plus-spring-boot3-starter</artifactId>
                 <version>${mybatis.plus.version}</version>
             </dependency>
             <dependency>
@@ -187,8 +199,8 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.8.1</version>
                 <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
+                    <source>17</source>
+                    <target>17</target>
                     <encoding>UTF-8</encoding>
                 </configuration>
             </plugin>
