@@ -1,6 +1,6 @@
 package com.mrqinzh.framework.common.web.utils;
 
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 
 public class WebFrameworkUtils {
 
