@@ -2,11 +2,10 @@ package com.mrqinzh.comment.domain.vo;
 
 import com.mrqinzh.framework.common.domain.vo.VO;
 import io.swagger.v3.oas.annotations.media.Schema;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
-
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 
 @Setter
 @Getter
