@@ -2,9 +2,8 @@ package com.mrqinzh.comment.rpc;
 
 import com.mrqinzh.comment.api.CommentApi;
 import com.mrqinzh.comment.service.CommentService;
+import jakarta.annotation.Resource;
 import org.springframework.web.bind.annotation.RestController;
-
-import javax.annotation.Resource;
 
 @RestController
 public class CommentServiceProvider implements CommentApi {
